fix: move routed experts helpers to utils
